fix: server

+++ server/src/dish-management/controllers/dish.controller.js
@@ -18,23 +18,26 @@ const getDishes = catchAsync(async (req, res) => {
 
 const createDish = catchAsync(async (req, res) => {
   const shopId = _.get(req, 'shop.id');
+  const userId = _.get(req, 'user.id');
   const createBody = req.body;
-  const dish = await dishService.createDish({ shopId, createBody });
+  const dish = await dishService.createDish({ shopId, createBody, userId });
   res.status(httpStatus.CREATED).send({ dish });
 });
 
 const updateDish = catchAsync(async (req, res) => {
   const shopId = _.get(req, 'shop.id');
   const dishId = _.get(req, 'params.dishId');
+  const userId = _.get(req, 'user.id');
   const updateBody = req.body;
-  await dishService.updateDish({ shopId, dishId, updateBody });
+  await dishService.updateDish({ shopId, dishId, updateBody, userId });
   res.status(httpStatus.OK).send({ message: 'Cập nhật thành công' });
 });
 
 const deleteDish = catchAsync(async (req, res) => {
   const shopId = _.get(req, 'shop.id');
   const dishId = _.get(req, 'params.dishId');
-  await dishService.deleteDish({ shopId, dishId });
+  const userId = _.get(req, 'user.id');
+  await dishService.deleteDish({ shopId, dishId, userId });
   res.status(httpStatus.OK).send({ message: 'Xoá thành công' });
 });
